Moved most component out of SeamMeme tsx
