+++ src/blocks/utils/SeamMeme/types/types.ts
@@ -1,0 +1,9 @@
+export interface Meme {
+    box_count: number;
+    captions: number;
+    height: number;
+    id: string;
+    name: string;
+    url: string;
+    width: string;
+}
